refactor: replace direct event setup with EventAggregatorTestHelper in PlayerSkillBonusesTests and PlayerSpeechTests
